+++ js/background/main.js
@@ -1,3 +1,4 @@
 import { initializeServiceWorker } from './background.js';
 
+console.log(`Spaces ${chrome.runtime.getManifest().version}`)
 initializeServiceWorker();