+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.8] - 2025-11-30
+
+### Changes
+
+- Fixed [issue #31](https://github.com/codedread/spaces/issues/31): Allow changing the capitalization of space names.
+- Increased unit test coverage from 16.17% to 22.69%.
+- 
 ## [1.1.7] - 2025-10-01
 
 ### Changes